+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Improvements
 - Only set a paragraph's area dirty when a text change would cause it to change size
 - Ensure that a panel gets notified of all relevant changes by calling OnChildAreaDirty for all grandchildren
 - Avoid unnecessary panel updates by using an Epsilon comparison when scrolling children
+- Allow setting a default text alignment for paragraphs in UiStyle
 
 Fixes
 - Fixed paragraph links having incorrect hover locations when using special text alignments